Fix type_union_implicit rule for Draft 0/1/2 to check schema# instead of hyper-schema#
The type_union_implicit rule was incorrectly checking for hyper-schema#
vocabularies for Draft 0, 1, and 2 instead of schema# vocabularies. This
caused the rule to add type arrays even when enum was already present,
which is incorrect behavior.

Fixed by changing the vocabulary checks from:
- http://json-schema.org/draft-02/hyper-schema#
- http://json-schema.org/draft-01/hyper-schema#
- http://json-schema.org/draft-00/hyper-schema#

To:
- http://json-schema.org/draft-02/schema#
- http://json-schema.org/draft-01/schema#
- http://json-schema.org/draft-00/schema#

Also updated the 6 affected test cases to have the correct expected values
(enum without type array).

Co-Authored-By: unknown <>